requirements.txt
# Core dependencies for the Personal Book Library Tracker
sqlalchemy>=2.0.0
textual>=0.41.0
python-dateutil>=2.8.0